Add 9.0-latest repository for python-influxdb installation
Python-influxdb is needed for Ceilometer in Mitaka. Because
this package provided by nightly xenial repository we add
and pin this repository with lower priority.

In this commit and mos9.0-proposed moved to
http://mirror.fuel-infra.org/mos-repos/ubuntu/snapshots/9.0-latest/dists/mos9.0-proposed/

Because mos9.0-proposed has not been released, priority of this
repository changed to 400 (less than default) and we'll install
only missed packages from it.